requirements.txt
streamlit>=1.28.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
pydub>=0.25.1
requests>=2.31.0
Pillow>=10.0.0 